LRDOCS-3716 Refactor service reference samples into blade.custom.service, blade.greedy.service.reference, and blade.reluctant.service.reference. Add README files too.
