+++ liferay-gradle/blade.custom.service/bnd.bnd
@@ -1,0 +1,1 @@
+Bundle-Version: 1.0.0
