+++ liferay-gradle/blade.custom.service/configs/com.liferay.blade.reluctant.service.reference.portlet.ReluctantServiceReferencePortlet.config
@@ -1,0 +1,1 @@
+_someService.target=(component.name=com.liferay.blade.custom.service.CustomServiceImpl)
