+++ liferay-gradle/blade.greedy.service.reference/bnd.bnd
@@ -1,0 +1,1 @@
+Bundle-Version: 1.0.0
