+++ liferay-gradle/blade.greedy.service.reference/src/main/java/com/liferay/blade/greedy/service/reference/impl/SomeServiceImpl.java
@@ -1,0 +1,34 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.blade.greedy.service.reference.impl;
+
+import com.liferay.blade.reluctant.service.reference.api.SomeService;
+
+import org.osgi.service.component.annotations.Component;
+
+/**
+ * @author James Hinkey
+ */
+@Component(immediate = true, service = SomeService.class)
+public class SomeServiceImpl implements SomeService {
+
+	@Override
+	public String doSomething() {
+		Class<?> clazz = getClass();
+
+		return clazz.getName();
+	}
+
+}
