--- liferay-gradle/blade.reluctant.service.reference.override/blade.reluctant.service.reference.custom/bnd.bnd
@@ -1,3 +1,0 @@
-Bundle-Name: blade.reluctant.service.reference.new
-Bundle-SymbolicName: blade.reluctant.service.reference.new
-Bundle-Version: 1.0.0
