--- liferay-gradle/blade.reluctant.service.reference.override/blade.reluctant.service.reference.custom/build.gradle
@@ -1,37 +1,0 @@
-repositories {
-flatDir {
-	   dirs 'lib'
-}
-}
-
-buildscript {
-	dependencies {
-		classpath group: "com.liferay", name: "com.liferay.gradle.plugins", version: "3.1.3"
-	}
-
-	repositories {
-
-		mavenLocal()
-
-		maven {
-			url "https://cdn.lfrs.sl/repository.liferay.com/nexus/content/groups/public"
-		}
-	}
-}
-
-apply plugin: "com.liferay.plugin"
-
-dependencies {
-	compile name: "com.liferay.blade.reluctant.service.reference.old", version: "1.0.0"
-
-	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "2.0.0"
-	compileOnly group: "org.osgi", name: "osgi.cmpn", version: "6.0.0"
-}
-
-repositories {
-	mavenLocal()
-
-	maven {
-		url "https://cdn.lfrs.sl/repository.liferay.com/nexus/content/groups/public"
-	}
-}
