--- liferay-gradle/blade.reluctant.service.reference.override/blade.reluctant.service.reference.custom/configs/com.liferay.blade.reluctant.service.reference.old.portlet.OverrideMyServiceReferencePortlet.cfg
@@ -1,1 +1,0 @@
-_someService.target=(component.name=com.liferay.blade.reluctant.service.reference.custom.CustomServiceImpl)
