--- liferay-gradle/blade.reluctant.service.reference.override/blade.reluctant.service.reference.old/bnd.bnd
@@ -1,4 +1,0 @@
-Bundle-Name: blade.reluctant.service.reference.old
-Bundle-SymbolicName: com.liferay.blade.reluctant.service.reference.old
-Bundle-Version: 1.0.0
-Export-Package: com.liferay.blade.reluctant.service.reference.old.service.api
