--- liferay-gradle/blade.reluctant.service.reference.override/blade.reluctant.service.reference.old/src/main/java/com/liferay/blade/reluctant/service/reference/old/service/api/SomeService.java
@@ -1,10 +1,0 @@
-package com.liferay.blade.reluctant.service.reference.old.service.api;
-
-/**
- * @author James Hinkey
- */
-public interface SomeService {
-
-	public String doSomething();
-
-}
