+++ liferay-gradle/blade.reluctant.service.reference/bnd.bnd
@@ -1,0 +1,4 @@
+Bundle-Name: blade.reluctant.service.reference
+Bundle-SymbolicName: com.liferay.blade.reluctant.service.reference
+Bundle-Version: 1.0.0
+Export-Package: com.liferay.blade.reluctant.service.reference.api
