+++ liferay-gradle/settings.gradle
@@ -5,10 +5,12 @@ include(
 	'blade.configurationaction',
 	'blade.controlmenuentry',
 	'blade.corejsphook',
+	'blade.custom.service',
 	'blade.doclib.resourcecommand.override',
 	'blade.document.action',
 	'blade.friendlyurl',
 	'blade.gogo',
+	'blade.greedy.service.reference',
 	'blade.hook.jsp',
 	'blade.hook.resourcebundle',
 	'blade.indexerpostprocessor',
@@ -32,6 +34,7 @@ include(
 	'blade.portlet.toolbar.contributor',
 	'blade.resourcebundle',
 	'blade.rest',
+	'blade.reluctant.service.reference',
 	'blade.schedulerentry',
 	'blade.screenname.validator',
 	'blade.service.hook.user',
